refactor(dashboard): チャプター9の内容をコミット
Streaming SSRについて学んだ。
次はPPRについて学ぶ。めっちゃ楽しみ。
