+++ app/dashboard/(overview)/page.tsx
@@ -1,0 +1,31 @@
+import { lusitana } from '@/app/ui/fonts';
+import { fetchCardData } from '../../lib/data';
+import RevenueChart from '../../ui/dashboard/revenue-chart';
+import CardWrapper from '../../ui/dashboard/cards';
+import { Suspense } from 'react';
+import { CardsSkeleton, LatestInvoicesSkeleton, RevenueChartSkeleton } from '@/app/ui/skeletons';
+import LatestInvoices from '@/app/ui/dashboard/latest-invoices';
+
+export default async function Page() {
+  const { numberOfInvoices, numberOfCustomers, totalPendingInvoices, totalPaidInvoices} = await fetchCardData();
+  return (
+    <main>
+      <h1 className={`${lusitana.className} mb-4 text-xl md:text-2xl`}>
+        Dashboard
+      </h1>
+      <div className="grid gap-6 sm:grid-cols-2 lg:grid-cols-4">
+        <Suspense fallback={<CardsSkeleton />}>
+          <CardWrapper />
+        </Suspense>
+      </div>
+      <div className="mt-6 grid grid-cols-1 gap-6 md:grid-cols-4 lg:grid-cols-8">
+        <Suspense fallback={<RevenueChartSkeleton />}>
+          <RevenueChart />
+        </Suspense>
+        <Suspense fallback={<LatestInvoicesSkeleton />}>
+          <LatestInvoices />
+        </Suspense>
+      </div>
+    </main>
+  );
+}
